refactor: extract code block module
